+++ code/src/utils/constants.ts
@@ -1,6 +1,9 @@
 export const XSD_DATETIME: string = "http://www.w3.org/2001/XMLSchema#dateTime";
 export const XSD_DATE: string = "http://www.w3.org/2001/XMLSchema#date";
 export const XSD_TIME: string = "http://www.w3.org/2001/XMLSchema#time";
+export const XSD_DECIMAL: string = "http://www.w3.org/2001/XMLSchema#decimal";
+export const XSD_INTEGER: string = "http://www.w3.org/2001/XMLSchema#integer";
+export const XSD_STRING: string = "http://www.w3.org/2001/XMLSchema#string";
 
 export const BULK_IDENTIFIER: string = "bulk";
 export const DATE_KEY: string = "date";